Beginning of modifications for cache optimizations

+++ app/src/main/java/com/github/brugapp/brug/view_model/ItemsMenuViewModel.kt
@@ -1,9 +1,10 @@
 package com.github.brugapp.brug.view_model
 
 import android.graphics.drawable.Drawable
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.liveData
+import androidx.lifecycle.viewModelScope
 import androidx.recyclerview.widget.RecyclerView
 import com.github.brugapp.brug.R
 import com.github.brugapp.brug.data.ItemsRepository
@@ -12,8 +13,10 @@ import com.github.brugapp.brug.ui.ITEMS_DELETE_TEXT
 import com.google.android.material.snackbar.Snackbar
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
-import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 
+
+private const val ITEMS_ERROR_TEXT = "ERROR: Unable to save your changes remotely. Try again later."
 
 /**
  * ViewModel of the Items Menu UI, handling its UI logic.
@@ -22,6 +25,7 @@ class ItemsMenuViewModel : ViewModel() {
 
     fun setCallback(
         activity: AppCompatActivity,
+        isTest: Boolean,
         dragPair: Pair<Int, Int>,
         swipePair: Pair<Drawable, Int>,
         listAdapterPair: Pair<MutableList<MyItem>, ItemsListAdapter>,
@@ -37,51 +41,43 @@ class ItemsMenuViewModel : ViewModel() {
             swipePair,
             listAdapterPair
         ) { deletedItem, position ->
-            liveData(Dispatchers.IO) {
-                emit(
-                    ItemsRepository.deleteItemFromUser(
-                        deletedItem.getItemID(), firebaseAuth.currentUser!!.uid, firestore
-                    )
-                )
-            }.observe(activity) { response ->
+            if(!isTest){
+                viewModelScope.launch {
+                    val result = ItemsRepository.deleteItemFromUser(
+                        deletedItem.getItemID(),
+                        firebaseAuth.currentUser!!.uid,
+                        firestore
+                    ).onSuccess
 
-                if (response.onSuccess) {
-                    listAdapterPair.first.removeAt(position)
-                    listAdapterPair.second.notifyItemRemoved(position)
-
-                    Snackbar.make(
-                        listView,
-                        ITEMS_DELETE_TEXT,
-                        Snackbar.LENGTH_LONG
-                    ).setAction("Undo") {
-                        liveData(Dispatchers.IO) {
-                            emit(
-                                ItemsRepository.addItemToUser(
-                                    deletedItem,
-                                    firebaseAuth.currentUser!!.uid, firestore
-                                )
-                            )
-                        }.observe(activity) {
-                            if (!response.onSuccess) {
-                                Snackbar.make(
-                                    listView,
-                                    "ERROR: Unable to re-add requested item to database",
-                                    Snackbar.LENGTH_LONG
-                                ).show()
-                            }
-                        }
-
-                        listAdapterPair.first.add(position, deletedItem)
-                        listAdapterPair.second.notifyItemInserted(position)
-                    }.show()
-                } else {
-                    Snackbar.make(
-                        listView,
-                        "ERROR: Unable to delete requested item from database",
-                        Snackbar.LENGTH_LONG
-                    ).show()
+                    if(!result){
+                        Toast.makeText(activity, ITEMS_ERROR_TEXT, Toast.LENGTH_LONG).show()
+                    }
                 }
             }
+            listAdapterPair.first.removeAt(position)
+            listAdapterPair.second.notifyItemRemoved(position)
+
+            Snackbar.make(
+                listView,
+                ITEMS_DELETE_TEXT,
+                Snackbar.LENGTH_LONG
+            ).setAction("Undo") {
+                if(!isTest) {
+                    viewModelScope.launch{
+                        val result = ItemsRepository.addItemToUser(
+                            deletedItem,
+                            firebaseAuth.currentUser!!.uid,
+                            firestore
+                        ).onSuccess
+
+                        if(!result){
+                            Toast.makeText(activity, ITEMS_ERROR_TEXT, Toast.LENGTH_LONG).show()
+                        }
+                    }
+                }
+                listAdapterPair.first.add(position, deletedItem)
+                listAdapterPair.second.notifyItemInserted(position)
+            }.show()
         }
     }
 }
